pyproject.toml
[tool.poetry]
name = "car_ai_vendor"
version = "0.1.0"
description = "An AI vendor for car data"
authors = ["domgjntoki"]
readme = "README.md"
license = "UNLICENSED"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes `E` and `F` codes by default.
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I001",
]
ignore = []

per-file-ignores = {}

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
httpx = "^0.27.0"
pydantic = {extras = ["email"], version = "^2.7.4"}
pydantic-settings = "^2.3.2"
asyncpg = "^0.29.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sentry-sdk = "^2.5.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
python-dotenv = "^1.1.0"
openai = "^1.76.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
faker = "^37.1.0"

[tool.poetry.group.prod.dependencies]
gunicorn = "^22.0.0"
python-json-logger = "^2.0.7"
prometheus-client = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"